+++ security/provenance/include/provenance_cgroup.h
@@ -6,8 +6,8 @@
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2, as
- * published by the Free Software Foundation; either version 2 of the License, or
- *	(at your option) any later version.
+ * published by the Free Software Foundation; either version 2 of the License,
+ * or (at your option) any later version.
  *
  */
 #ifndef CONFIG_SECURITY_PROVENANCE_CGROUP
@@ -64,5 +64,4 @@ static inline uint8_t prov_cgroup_add_or_update(struct cgroup_filters *f)
 	list_add_tail(&(f->list), &cgroup_filters); // not already on the list, we add it
 	return 0;
 }
-
 #endif