+++ security/provenance/include/provenance_net.h
@@ -6,8 +6,8 @@
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2, as
- * published by the Free Software Foundation; either version 2 of the License, or
- *	(at your option) any later version.
+ * published by the Free Software Foundation; either version 2 of the License,
+ * or (at your option) any later version.
  *
  */
 #ifndef CONFIG_SECURITY_PROVENANCE_NET
@@ -35,7 +35,7 @@ static inline struct provenance *sk_inode_provenance(struct sock *sk)
 {
 	struct socket *sock = sk->sk_socket;
 
-	if (sock == NULL)
+	if (!sock)
 		return NULL;
 	return socket_inode_provenance(sock);
 }
@@ -50,9 +50,9 @@ static inline struct provenance *sk_provenance(struct sock *sk)
 #define ihlen(ih) (ih->ihl * 4)
 
 static inline void __extract_tcp_info(struct sk_buff *skb,
-																			struct iphdr *ih,
-																			int offset,
-																			struct packet_identifier *id)
+				      struct iphdr *ih,
+				      int offset,
+				      struct packet_identifier *id)
 {
 	struct tcphdr _tcph;
 	struct tcphdr *th;
@@ -70,12 +70,12 @@ static inline void __extract_tcp_info(struct sk_buff *skb,
 }
 
 static inline void __extract_udp_info(struct sk_buff *skb,
-																			struct iphdr *ih,
-																			int offset,
-																			struct packet_identifier *id)
+				      struct iphdr *ih,
+				      int offset,
+				      struct packet_identifier *id)
 {
 	struct udphdr _udph;
-	struct udphdr	*uh;
+	struct udphdr   *uh;
 	int udpoff;
 
 	if (ntohs(ih->frag_off) & IP_OFFSET)
@@ -88,7 +88,7 @@ static inline void __extract_udp_info(struct sk_buff *skb,
 	id->rcv_port = uh->dest;
 }
 
-static inline unsigned int provenance_parse_skb_ipv4(struct sk_buff *skb, union prov_msg *prov)
+static inline unsigned int provenance_parse_skb_ipv4(struct sk_buff *skb, union prov_elt *prov)
 {
 	struct packet_identifier *id;
 	int offset;
@@ -97,13 +97,13 @@ static inline unsigned int provenance_parse_skb_ipv4(struct sk_buff *skb, union 
 
 	offset = skb_network_offset(skb);
 	ih = skb_header_pointer(skb, offset, sizeof(_iph), &_iph); // we obtain the ip header
-	if (ih == NULL)
+	if (!ih)
 		return -EINVAL;
 
 	if (ihlen(ih) < sizeof(_iph))
 		return -EINVAL;
 
-	memset(prov, 0, sizeof(union prov_msg));
+	memset(prov, 0, sizeof(union prov_elt));
 	id = &packet_identifier(prov); // we are going fo fill this
 
 	id->type = ENT_PACKET;
@@ -190,39 +190,79 @@ static inline uint8_t prov_ipv4_add_or_update(struct list_head *filters, struct 
 }
 
 // incoming packet
-static inline void record_pck_to_inode(union prov_msg *pck, struct provenance *inode)
+static inline int record_pck_to_inode(union prov_elt *pck, struct provenance *inode)
 {
-	union prov_msg relation;
+	int rc = 0;
 
-	if (unlikely(pck == NULL || inode == NULL)) // should not occur
-		return;
-	if (!provenance_is_tracked(prov_msg(inode)) && !prov_all)
-		return;
-	if (!should_record_relation(RL_RCV_PACKET, pck, prov_msg(inode), FLOW_ALLOWED))
-		return;
-	memset(&relation, 0, sizeof(union prov_msg));
+	if (unlikely(!pck || !inode)) // should not occur
+		return 0;
+	if (!provenance_is_tracked(prov_elt(inode)) && !prov_all)
+		return 0;
+	if (!should_record_relation(RL_RCV_PACKET, pck, prov_elt(inode)))
+		return 0;
+	rc = __update_version(RL_RCV_PACKET, inode);
+	if (rc < 0)
+		return rc;
+	write_node(prov_elt(inode));
 	prov_write(pck);
-	__record_node(prov_msg(inode));
-	__update_version(RL_RCV_PACKET, inode);
-	__record_node(prov_msg(inode));
-	__record_relation(RL_RCV_PACKET, &(pck->msg_info.identifier), &(prov_msg(inode)->msg_info.identifier), &relation, FLOW_ALLOWED, NULL);
+	rc = write_relation(RL_RCV_PACKET, pck, prov_elt(inode), NULL);
+	return rc;
 }
 
 // outgoing packet
-static inline void record_inode_to_pck(struct provenance *inode, union prov_msg *pck)
+static inline int record_inode_to_pck(struct provenance *inode, union prov_elt *pck)
 {
-	union prov_msg relation;
+	int rc = 0;
 
-	if (unlikely(pck == NULL || inode == NULL)) // should not occur
-		return;
-	if (!provenance_is_tracked(prov_msg(inode)) && !prov_all)
-		return;
-	if (!should_record_relation(RL_SND_PACKET, prov_msg(inode), pck, FLOW_ALLOWED))
-		return;
-	memset(&relation, 0, sizeof(union prov_msg));
-	__record_node(prov_msg(inode));
+	if (unlikely(!pck || !inode)) // should not occur
+		return 0;
+	if (!provenance_is_tracked(prov_elt(inode)) && !prov_all)
+		return 0;
+	if (!should_record_relation(RL_SND_PACKET, prov_elt(inode), pck))
+		return 0;
+	write_node(prov_elt(inode));
 	prov_write(pck);
-	__record_relation(RL_SND_PACKET, &(prov_msg(inode)->msg_info.identifier), &(pck->msg_info.identifier), &relation, FLOW_ALLOWED, NULL);
+	rc = write_relation(RL_SND_PACKET, prov_elt(inode), pck, NULL);
 	inode->has_outgoing = true;
+	return rc;
+}
+
+static inline int provenance_record_address(struct sockaddr *address, int addrlen, struct provenance *prov)
+{
+	union long_prov_elt *addr_info;
+	int rc = 0;
+
+	if (provenance_is_name_recorded(prov_elt(prov)) || !provenance_is_recorded(prov_elt(prov)))
+		return 0;
+	addr_info = alloc_long_provenance(ENT_ADDR);
+	if (!addr_info) {
+		rc = -ENOMEM;
+		goto out;
+	}
+	addr_info->address_info.length = addrlen;
+	memcpy(&(addr_info->address_info.addr), address, addrlen);
+	write_long_node(addr_info);
+	rc = write_relation(RL_NAMED, addr_info, prov_elt(prov), NULL);
+	set_name_recorded(prov_elt(prov));
+out:
+	free_long_provenance(addr_info);
+	return rc;
+}
+
+static inline int record_packet_content(union prov_elt *pck, const struct sk_buff *skb)
+{
+	union long_prov_elt *cnt = alloc_long_provenance(ENT_PCKCNT);
+	int rc;
+
+	cnt->pckcnt_info.length = skb_end_offset(skb);
+	if (cnt->pckcnt_info.length > PATH_MAX) {
+		cnt->pckcnt_info.truncated = PROV_TRUNCATED;
+		memcpy(cnt->pckcnt_info.content, skb->head, PATH_MAX);
+	} else
+		memcpy(cnt->pckcnt_info.content, skb->head, cnt->pckcnt_info.length);
+	write_long_node(cnt);
+	rc = write_relation(RL_READ, cnt, pck, NULL);
+	free_long_provenance(cnt);
+	return rc;
 }
 #endif