+++ security/provenance/netfilter.c
@@ -6,13 +6,10 @@
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2, as
- * published by the Free Software Foundation; either version 2 of the License, or
- *	(at your option) any later version.
+ * published by the Free Software Foundation; either version 2 of the License,
+ * or (at your option) any later version.
  *
  */
-
-#if defined(CONFIG_NETFILTER)
-
 #include "provenance.h"
 #include "provenance_net.h"
 #include "provenance_task.h"
@@ -21,21 +18,21 @@ static inline unsigned int __ipv4_out(struct sk_buff *skb)
 {
 	struct provenance *cprov = current_provenance();
 	struct provenance *iprov = NULL;
-	union prov_msg pckprov;
+	union prov_elt pckprov;
 	unsigned long irqflags;
 
-	if (cprov == NULL)
+	if (!cprov)
 		return NF_ACCEPT;
 
-	if (provenance_is_tracked(prov_msg(cprov))) {
+	if (provenance_is_tracked(prov_elt(cprov))) {
 		iprov = sk_inode_provenance(skb->sk);
-		if (iprov == NULL)
+		if (!iprov)
 			return NF_ACCEPT;
 		provenance_parse_skb_ipv4(skb, &pckprov);
 		spin_lock_irqsave_nested(prov_lock(cprov), irqflags, PROVENANCE_LOCK_TASK);
 		spin_lock_nested(prov_lock(iprov), PROVENANCE_LOCK_INODE);
 		record_inode_to_pck(iprov, &pckprov);
-		if (provenance_records_packet(prov_msg(iprov)))
+		if (provenance_records_packet(prov_elt(iprov)))
 			record_packet_content(&pckprov, skb);
 		spin_unlock(prov_lock(iprov));
 		spin_unlock_irqrestore(prov_lock(cprov), irqflags);
@@ -67,9 +64,7 @@ static int __init provenance_nf_init(void)
 	err = nf_register_hooks(provenance_nf_ops, ARRAY_SIZE(provenance_nf_ops));
 	if (err)
 		panic("Provenance: nf_register_hooks: error %d\n", err);
-	printk(KERN_INFO "Provenance netfilter ready.\n");
+	pr_info("Provenance: netfilter ready.\n");
 	return 0;
 }
 module_init(provenance_nf_init);
-MODULE_LICENSE("GPL");
-#endif