+++ CHANGES.md
@@ -22,6 +22,7 @@
 
 ### v0.3.0
 ```
+- add a terminate relationship when a task is freed.
 - boot id is set from userspace.
 - union prov_msg -> union prov_elt
 - Implemented run-time query system.